refactor(plateform): refacto events logic
