+++ plateform/app/routes/quiz/age.tsx
@@ -2,7 +2,8 @@ import { useEffect, useState, type SubmitEvent } from "react";
 import { useLocation, useOutletContext } from "react-router";
 import Label from "~/components/quiz/label";
 import NextButton from "~/components/quiz/next-button";
-import { resolveQuizEntrySource, trackQuizStarted } from "~/services/tracking/events";
+import { trackQuizStarted } from "~/services/tracking/events";
+import { resolveQuizEntrySource } from "~/services/tracking/utils";
 import { useQuizStore } from "~/stores/quiz";
 import { isValidAge } from "~/utils/quiz";
 import type { QuizOutletContext } from "./_layout";
