+++ plateform/app/routes/results.tsx
@@ -17,7 +17,8 @@ import { OPTIONS } from "~/config/quiz-options";
 import { useIsMobile } from "~/hooks/useIsMobile";
 import { useMissionResults } from "~/hooks/useMissionResults";
 import { setQuizSessionId } from "~/services/tracking";
-import { trackMissionClickedFromMatch, trackResultsViewed, type MissionDetailNavState } from "~/services/tracking/events";
+import { trackMissionClickedFromMatch, trackResultsViewed } from "~/services/tracking/events";
+import type { MissionDetailNavState } from "~/services/tracking/types";
 import { useQuizStore } from "~/stores/quiz";
 import { evalCondition } from "~/utils/conditions";
 import { buildMissionDetailHref, matchResultToBrowseMission } from "~/utils/mission";
